+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+- 2.2.0 (2018/5/5)
+	- Support css color name and hex conversion.
+	- Support converting the CSS colorname in the entire file to hex.
+	- Fix text display bug.
+
 - 2.1.0 (2018/4/25)
   - add feature (`hex_convert_to_colorname`)